Fix NULL pointer difference in tpe module
I had thought this bug was a threading or locking issue, but it wasn't.
I finally did a bunch of straces to find out what exactly in userspace
was the last call before the error, turns out it happened when you tried
to execute something that didn't exist.

I had this:

	file = open_exec(name);

	if (file == NULL) {
		return -ENOENT;
	}

That is incorrect. File isn't null on error, it's an error code.

This is correct:

	if (IS_ERR(file))
		return file;

Ta da! No more bug.