Wrap security_bprm_check instead of do_execve
By wrapping just the security_ functions, which are only used if you use
a security framework anyway, we can completly bypass the "invalid opcode"
but that has been cropping up. In doing some resarch into which security
functions are used in the execve process, I found security_bprm_check.

Happy days.

The old behaviour can still be enabled by setting WRAP_SYSCALLS to 1. Do
so at your own risk, and probably only if you want to troubleshoot that
problem, as it would be interesting to find a stable way to do that. You
know, to easily write rootkits :)