Major rewrite of hijack code
I would like to give a big thanks to Eugene Shatokhin for his help with this
code. He sent me a copy of his detour module, which is not yet published, and
it contained the code (which now mostly resides in hijacks.c) to copy and fixup
functions.

This module now appers very stable, and functions are now wrapped properly
rather than hijacked completely. I'll have to play around with this more, but
it's great so far.

Thanks again, Eugene!