Only process the d_path() stuff when needed
Previously, we did a lookup of the parent, and the full path of the files of
the tasks, upon each and every execute. There is no neeed for that. We only
need this information when it's actually denied. So, move the logic to a new
function called log_denied_exec();