Add "check_file" option to TPE
Previously we only did the checking on the parent inode, ie, the directory the
file was in. While that is normally good enough, if for some reason there is a
file that a user can write to for another reason (they own it, group/world
writable, etc), deny the execution.

This may cause some programs to fail, particularly custom applications, that
suid as a non-root user. Because of this, a "check_file" option has been added
as a sysctl option.