Fix DoS condition on recursive logging
This patch fixes a denial-of-service condition when exec is denied on a process
with a lot of parents. It's caused by the parent_task_walk() being a recursive
function.

Recursive function in the linux kernel? What was I thinking? I wasn't,
actually, as my main goal with the project initially was proof-of-concept. Now
that I'm actually using the thing in ad-hoc production environments, it's time
I search the code for problems. Well, I found one!

This patch also introduces a new sysctl entry: log_max