Implement "lock" for tpe sysctl entries
When the lock flag is enabled, you won't be able to change any of the sysctl
entries for this module. Note that the module can just be unloaded/loaded. If
you enable the lock, you will also want to enable this:

/proc/sys/kernel/modules_disabled

If your kernel doesn't have modules_disable (such as the el5 guys), sucks to be
you. Maybe in the future I'll implement one with this module.