Fix workaround in "ps" feature
While you couldn't read the /proc/<pid>/ directory as a non-root user, you
could still read the contents of /proc/<pid>/* because procfs isn't a normal
filesystem. So you could still view other processes if you wrote a custom tool
to do it. The snmpd also had no issue seeing other processes.

Luckily, the pid_revalidate() symbol is called when reading those files. So
hijack that to complete the "ps" feature and stop this workaround. Hopefully
that solved the problem for good :)