Add find_address_symbol() function
This finds a symbol name based on the given address, which is the reverse of
what the find_symbol_address() function does. It's helping me debug some
issues with some features I'm trying to add, and it looks like I'll actually be
keeping it around anyway so might as well commit it :)

Currently this is only for kernel versions >= 2.6.30, I'll eventually code in
the matching code for previous versions once it's actually being used in the
main code flow.

Note that I'm changing sym->name from a const char * to a char * (and adding
the name_alloc boolean), this is because the kallsyms callback has the function
name on the stack, and goes missing once outside that scope. So we need to
allocate memory to copy it to, and remember we did that so we can later free
it.