Verify that mm_struct * isn't NULL in exe_from_mm
So another oops on my part. Not checking if the argument is null.

I triggered this by chown'ing an application in my Ubuntu Unity panel launcher
to my user and clicking on it. Boy was that a nasty suprise. The parent has no
task->mm, not sure why. Here's the deny message now that this is fixed:

  [tpe] Denied untrusted exec of /usr/bin/gcalctool (uid:1000) by <d_path
  failed> (uid:1000), parents: /usr/bin/compiz (uid:1000),
  /usr/bin/gnome-session (uid:1000), /usr/sbin/lightdm (uid:0),
  /usr/sbin/lightdm (uid:0), /sbin/init (uid:0). Deny reason: file uid not
  trusted

I'm _very_ curious as to why this is the case. I'll dig in some more later.

2nd NULL pointer found so far. I really ought to be more careful! :)