Fix build issues on newer kernels (CentOS 7 & Ubuntu 14)
A rather large commit that fixes several issues, I know, but I'm only starting to get back into the developer swing of things :)

The following changes are in this commit:

* malloc.c was renamed to kernfunc.c, since it now handles more than just finding the malloc functions in the kernel
* kernfunc.c now resolves the symbols for inat and insn, which means we no longer need the arch/x86 code
* compilation errors due to the introduction of CONFIG_UIDGID_STRICT_TYPE_CHECKS have been resolved
* rename of VM_EXECUTABLE to VM_EXEC in newer kernels handled
* FS_REVAL_DOT ABI change to use d_revalidate handled
* generic_permission ABI change handled
* security_file_mmap ABI change handled
* security_task_setuid ABI change handled
* setuid protection no longer applies to users in the trusted_gid

NOTE TO EL6 AND EL5 USERS (if you're bothering to read the change log, that is) - I'm repeating the last line for your benefit. The setuid protection option now behaves slightly different. It no longer applies to users in the trusted_gid. There is no need for you to upgrade to this release if this bothers you for some reason. It shouldn't, but who knows?

This commit should add full support for RHEL/CentOS 7 (linux-3.10.0) and Ubuntu 14.04 LTS (linux-3.13.0). If there are build failures on other versions of kernels less than these, let me know, probably an easy fix.

This commit does not address compile time issues in EL5. I'll address those sometime later.

Finally, I haven't done a full regression test of this. I'll be committing some more files ot the tests/ directory to do this.