Initial commit of a basic test suite
The goal here is to exercise each feature to make sure it works, and that the system doesn't crash in the process ;) Not every test is fully implemented (they will say SKIP instead of PASS/FAIL) and I'm not sure the mmap test is working properly.

The tests are sort of unconventional, I know, fiddling with stuff directly in /bin ... but that's part of the trusted path. What else am I supposed to do?