Fix mmap protection regression in EL7
More shenanigans with ABI changes. The security_file_mmap function was renamed to security_mmap_file with the addition of security_mmap_addr, but neither are correctly placed in the path of do_mmap_pgoff. This is a problem in the linux kernel, but of course, nobody noticed, and I have to wonder if anyone is going to care. I'd rather hook into the LSM function, but since this is a kernel issue that I can't get around, I'll just hook directly into do_mmap_pgoff on all kernels and be done with it. It's what I did on EL5 anyway.