Reimplement TPE using ftrace
Because I honestly don't understand the insn portion of this code, I am
reimplementing it with ftrace because 1) I know what it's doing and 2)
it'll be a lot easier for me to maintain. Newer kernels were faulting
on the insn section, and rather than headache after headache trying to
debug it, just use ftrace instead.

Now, we're regressing on functionality (ie; getting rid of the "extras"
and the sysctl lock). This is because I don't know how to properly set
CPU registers in the ftrace'd function and return to the INSN offset
address properly (see, that pesky insn stuff again?). So, until someone
helps me with that, I am slimming down this module to just the core TPE
functionality.

So far only tested on EL7. EL6 system continue to work with the old
code, so use that if you're still on it.

There's currently one outstanding issue, this happens on insert:

WARNING: at kernel/trace/ftrace.c:440 ftrace_startup+0x288/0x2e0()

I'm looking into it :)