Remove a layer of indirection and cleanup some code
By using the pt_regs directly from the handler, we can preserve the
regs on the calling function and decide whether or not to return to
it from the handler itself.

This may open the door back up to implement all the "extras" that
were removed in the last commit. More testing required.