explicity require selinux to be loaded for cred remap
The chosen_lsm symbol exists in the code section because it's tagged
with __initdata, so that makes searching for it via kallsyms return
jibberish. So, we can't use it. BOOO!!

While this module works with LSM disabled, there's no easy way to
tell whether or not it is in-fact disabled. Could use the security_ops
struct, but that no longer exists in newer kernels. Again, BOOO!!

Most people disable selinux with the config file, which means it's
loaded just not enabled. We can detect that this way. Good enough.